types: Migrate all of the Info types
